Generate TLEs from ephemeris data
Move ephemeris data collection out of satellite_utils, generate TLE from best fit of ephemeris positions, add benchmark test for ephemeris data values
