+++ src/data/satellite_utils.py
@@ -1,25 +1,13 @@
 import logging
 import re
 import time
-from datetime import datetime, timedelta, timezone
+from datetime import datetime, timezone
 
-import numpy as np
 import pandas as pd
 import psycopg2
 import requests
 from bs4 import BeautifulSoup
 from connections import get_spacetrack_login
-from psycopg2.extras import execute_values
-
-from api.domain.models.interpolable_ephemeris import (
-    EphemerisPoint,
-    InterpolableEphemeris,
-)
-from api.domain.models.satellite import Satellite
-from api.utils.interpolation_utils import (
-    generate_and_propagate_sigma_points,
-    interpolate_sigma_pointsKI,
-)
 
 
 def get_decayed_satellites(cursor, connection):
@@ -334,534 +322,3 @@ def get_starlink_generations(cursor, connection):
     logging.info(
         f"Updated generation information for {total_matched} Starlink satellites"
     )
-
-
-def insert_ephemeris_data(
-    parsed_data, cursor, connection
-) -> tuple[InterpolableEphemeris, int] | None:
-    """
-    Insert ephemeris data into the database efficiently using batch inserts.
-
-    Args:
-        parsed_data (dict): Dictionary containing parsed ephemeris data
-        cursor: Database cursor
-        connection: Database connection
-
-    Returns:
-        InterpolableEphemeris: The created ephemeris object if successful,
-        None if duplicate
-    """
-    try:
-        # TODO: fix usage of has_current_sat_number
-        # Get the satellite data to create a proper Satellite object
-        cursor.execute(
-            "SELECT id, sat_number, sat_name, constellation, generation, "
-            "rcs_size, launch_date, decay_date, object_id, object_type, "
-            "has_current_sat_number FROM satellites "
-            "WHERE sat_name = %s "
-            "AND has_current_sat_number = true",
-            (parsed_data["satellite_name"],),
-        )
-        satellite_result = cursor.fetchone()
-        if satellite_result is None:
-            logging.error(f"Satellite not found: {parsed_data['satellite_name']}")
-            return None
-
-        satellite_id = satellite_result[0]
-
-        date_collected = datetime.now(timezone.utc)
-        data_source = "starlink"
-
-        ephemeris_insert = """
-            INSERT INTO interpolable_ephemeris (
-                satellite,
-                date_collected,
-                generated_at,
-                data_source,
-                file_reference,
-                ephemeris_start,
-                ephemeris_stop,
-                frame
-            ) VALUES (
-                %s,
-                %s,
-                %s,
-                %s,
-                %s,
-                %s,
-                %s,
-                %s
-            )
-            ON CONFLICT (satellite, generated_at, data_source) DO NOTHING
-            RETURNING id
-        """
-
-        cursor.execute(
-            ephemeris_insert,
-            (
-                satellite_id,
-                date_collected,
-                parsed_data["generated_at"],
-                data_source,
-                parsed_data["filename"],
-                parsed_data["ephemeris_start"],
-                parsed_data["ephemeris_stop"],
-                parsed_data["frame"],
-            ),
-        )
-        ephemeris_result = cursor.fetchone()
-        if ephemeris_result is None:
-            logging.error(
-                f"Likely duplicate ephemeris data for {parsed_data['satellite_name']}"
-            )
-            return None
-        ephemeris_id = ephemeris_result[0]
-
-        # Prepare the points data for batch insert
-        timestamps = parsed_data["timestamps"]
-        positions = parsed_data["positions"]
-        velocities = parsed_data["velocities"]
-        covariances = parsed_data["covariances"]
-
-        # Create a list of tuples for batch insert
-        points = []
-        for i in range(len(timestamps)):
-            # Position and velocity are 3D arrays (x,y,z)
-            pos_array = [float(x) for x in positions[i]]
-            vel_array = [float(x) for x in velocities[i]]
-            # Covariance is a 6x6 matrix flattened to 1D array of 36 elements
-            cov_array = [float(x) for x in covariances[i].flatten()]
-
-            points.append(
-                (ephemeris_id, timestamps[i], pos_array, vel_array, cov_array)
-            )
-
-        # Batch insert the points with ON CONFLICT DO NOTHING
-        points_insert = """
-            INSERT INTO ephemeris_points (
-                ephemeris_id,
-                timestamp,
-                position,
-                velocity,
-                covariance
-            ) VALUES %s
-            ON CONFLICT (ephemeris_id, timestamp) DO NOTHING
-        """
-        execute_values(cursor, points_insert, points)
-
-        # Commit the transaction
-        connection.commit()
-
-        logging.info(
-            f"Successfully inserted {len(points)} "
-            f"points for ephemeris_id {ephemeris_id}"
-        )
-
-        # Create EphemerisPoint objects from the data
-        ephemeris_points = []
-        for i in range(len(timestamps)):
-            point = EphemerisPoint(
-                timestamp=timestamps[i],
-                position=np.array(positions[i]),
-                velocity=np.array(velocities[i]),
-                covariance=np.array(covariances[i]),
-            )
-            ephemeris_points.append(point)
-
-        # Create Satellite object from database result
-        satellite = Satellite(
-            sat_number=satellite_result[1],
-            sat_name=satellite_result[2],
-            constellation=satellite_result[3],
-            generation=satellite_result[4],
-            rcs_size=satellite_result[5],
-            launch_date=satellite_result[6],
-            decay_date=satellite_result[7],
-            object_id=satellite_result[8],
-            object_type=satellite_result[9],
-            has_current_sat_number=satellite_result[10],
-        )
-        # Create and return the InterpolableEphemeris object
-        ephemeris = InterpolableEphemeris(
-            satellite=satellite,
-            generated_at=parsed_data["generated_at"],
-            data_source=data_source,
-            frame=parsed_data["frame"],
-            points=ephemeris_points,
-            ephemeris_start=parsed_data["ephemeris_start"],
-            ephemeris_stop=parsed_data["ephemeris_stop"],
-            file_reference=parsed_data["filename"],
-            date_collected=date_collected,
-        )
-
-        return ephemeris, ephemeris_id
-
-    except Exception as e:
-        connection.rollback()
-        logging.error(f"Error inserting ephemeris data: {e}")
-        raise
-
-
-def parse_ephemeris_file(file_content: str, filename: str) -> dict:
-    """
-    Parse a satellite ephemeris file in UVW frame format.
-
-    Args:
-        file_content (str): Content of the ephemeris file
-        filename (str): Name of the file, used to extract satellite name
-
-    Returns:
-        dict: A dictionary containing:
-            - headers (dict): Key-value pairs from the file header
-            - timestamps (np.ndarray): Array of datetime objects
-            - positions (np.ndarray): Array of position vectors
-            - velocities (np.ndarray): Array of velocity vectors
-            - covariances (np.ndarray): Array of 6x6 covariance matrices
-            - frame (str): Coordinate frame (UVW)
-            - ephemeris_start (datetime): Start time of the ephemeris
-            - ephemeris_stop (datetime): End time of the ephemeris
-            - generated_at (datetime): When the ephemeris was created
-            - satellite_name (str): Name of the satellite extracted from filename
-    """
-    lines = file_content.splitlines()
-
-    # Parse headers
-    headers = {}
-    for i in range(3):
-        line = lines[i].strip()
-
-        if "ephemeris_start:" in line and "ephemeris_stop:" in line:
-
-            # everything after "ephemeris_start:" until "ephemeris_stop:"
-            start_match = re.search(
-                r"ephemeris_start:([^e]+?)(?=ephemeris_stop:)", line
-            )
-            if start_match:
-                headers["ephemeris_start"] = start_match.group(1).strip()
-
-            # everything after "ephemeris_stop:" until "step_size:"
-            stop_match = re.search(r"ephemeris_stop:([^s]+?)(?=step_size:)", line)
-            if stop_match:
-                headers["ephemeris_stop"] = stop_match.group(1).strip()
-
-            # everything after "step_size:"
-            step_match = re.search(r"step_size:(.+)", line)
-            if step_match:
-                headers["step_size"] = step_match.group(1).strip()
-        else:
-            if ":" in line:
-                colon_pos = line.find(":")
-                key = line[:colon_pos]
-                value = line[colon_pos + 1 :].strip()
-                headers[key] = value
-
-    if lines[3].strip() != "UVW":
-        raise ValueError("Expected UVW frame specification")
-
-    # Extract satellite name from filename
-    # Format: MEME_57851_STARLINK-30405_1490204_Operational_1432778700_UNCLASSIFIED
-    satellite_name = None
-    try:
-        # Find STARLINK- pattern and extract the full satellite name
-        if "STARLINK-" in filename:
-            start_idx = filename.find("STARLINK-")
-            end_idx = filename.find("_", start_idx)
-            if end_idx == -1:
-                end_idx = len(filename)
-            satellite_name = filename[start_idx:end_idx]
-    except Exception as e:
-        logging.warning(f"Could not parse satellite name from filename {filename}: {e}")
-
-    # Parse creation time
-    generated_at = None
-    if "created" in headers:
-        try:
-            time_str = headers["created"].replace(" UTC", "")
-            # Try parsing with time first, then fall back to date only
-            try:
-                generated_at = datetime.strptime(time_str, "%Y-%m-%d %H:%M:%S")
-            except ValueError:
-                generated_at = datetime.strptime(time_str, "%Y-%m-%d")
-            generated_at = generated_at.replace(tzinfo=timezone.utc)
-        except ValueError:
-            logging.warning(f"Could not parse creation time: {headers['created']}")
-
-    # Parse ephemeris start and stop times
-    ephemeris_start = None
-    ephemeris_stop = None
-    if "ephemeris_start" in headers and "ephemeris_stop" in headers:
-        try:
-            start_str = headers["ephemeris_start"].replace(" UTC", "")
-            stop_str = headers["ephemeris_stop"].replace(" UTC", "")
-            # Try parsing with time first, then fall back to date only
-            try:
-                ephemeris_start = datetime.strptime(start_str, "%Y-%m-%d %H:%M:%S")
-                ephemeris_stop = datetime.strptime(stop_str, "%Y-%m-%d %H:%M:%S")
-            except ValueError:
-                ephemeris_start = datetime.strptime(start_str, "%Y-%m-%d")
-                ephemeris_stop = datetime.strptime(stop_str, "%Y-%m-%d")
-            ephemeris_start = ephemeris_start.replace(tzinfo=timezone.utc)
-            ephemeris_stop = ephemeris_stop.replace(tzinfo=timezone.utc)
-        except ValueError:
-            logging.warning(
-                f"Could not parse ephemeris times: "
-                f"{headers['ephemeris_start']} - "
-                f"{headers['ephemeris_stop']}"
-            )
-
-    timestamps = []
-    positions = []
-    velocities = []
-    covariances = []
-
-    i = 4
-    while i < len(lines):
-        state_parts = lines[i].strip().split()
-
-        # Parse timestamp
-        timestamp_str = state_parts[0]
-        year = int(timestamp_str[:4])
-        day_of_year = int(timestamp_str[4:7])
-        hour = int(timestamp_str[7:9])
-        minute = int(timestamp_str[9:11])
-        second = int(timestamp_str[11:13])
-        # Handle decimal seconds
-        millisec = int(
-            float(timestamp_str[13:]) * 1000
-        )  # Convert decimal seconds to milliseconds
-
-        timestamp = datetime(year, 1, 1, tzinfo=timezone.utc).replace(
-            hour=hour, minute=minute, second=second, microsecond=millisec * 1000
-        ) + timedelta(days=day_of_year - 1)
-
-        # Parse position and velocity
-        pos = np.array([np.float64(x) for x in state_parts[1:4]])
-        vel = np.array([np.float64(x) for x in state_parts[4:7]])
-
-        # Parse covariance
-        cov_values = []
-        for j in range(3):
-            cov_values.extend([np.float64(x) for x in lines[i + 1 + j].strip().split()])
-
-        cov_matrix = np.zeros((6, 6), dtype=np.float64)
-        idx = 0
-        for row in range(6):
-            for col in range(row + 1):
-                cov_matrix[row, col] = cov_values[idx]
-                cov_matrix[col, row] = cov_values[idx]
-                idx += 1
-
-        # Store arrays
-        timestamps.append(timestamp)
-        positions.append(pos)
-        velocities.append(vel)
-        covariances.append(cov_matrix)
-
-        i += 4
-
-    # If we couldn't parse from headers, use first and last timestamps
-    if ephemeris_start is None:
-        ephemeris_start = timestamps[0]
-    if ephemeris_stop is None:
-        ephemeris_stop = timestamps[-1]
-    if generated_at is None:
-        generated_at = ephemeris_start
-
-    # Filter data to only include 26 hours (1 day + 2 hours) from start
-    timestamps_array = np.array(timestamps)
-    positions_array = np.array(positions, dtype=np.float64)
-    velocities_array = np.array(velocities, dtype=np.float64)
-    covariances_array = np.array(covariances, dtype=np.float64)
-
-    # Calculate cutoff time (26 hours from ephemeris start)
-    cutoff_time = ephemeris_start + timedelta(hours=26)
-
-    time_mask = timestamps_array <= cutoff_time
-
-    # Filter all arrays using the mask
-    filtered_timestamps = timestamps_array[time_mask]
-    filtered_positions = positions_array[time_mask]
-    filtered_velocities = velocities_array[time_mask]
-    filtered_covariances = covariances_array[time_mask]
-
-    # Update ephemeris_stop to reflect the actual end time of filtered data
-    if len(filtered_timestamps) > 0:
-        actual_ephemeris_stop = filtered_timestamps[-1]
-        ephemeris_stop = min(cutoff_time, actual_ephemeris_stop)
-    else:
-        ephemeris_stop = cutoff_time
-
-    original_count = len(timestamps)
-    filtered_count = len(filtered_timestamps)
-    logging.info(f"Original data points: {original_count}")
-    logging.info(f"Filtered data points (26 hours): {filtered_count}")
-    logging.info(f"Data reduction: {original_count - filtered_count} points removed")
-    logging.info(f"Parsed {filtered_count} timestamps")
-    logging.info(f"Parsed {len(filtered_positions)} positions")
-    logging.info(f"Parsed {len(filtered_velocities)} velocities")
-    logging.info(f"Parsed {len(filtered_covariances)} covariances")
-    logging.info(f"Ephemeris start: {ephemeris_start}")
-    logging.info(f"Ephemeris stop: {ephemeris_stop}")
-    logging.info(f"Generated at: {generated_at}")
-    logging.info(f"Satellite name: {satellite_name}")
-    logging.info(f"Filename: {filename}")
-
-    return {
-        "timestamps": filtered_timestamps,
-        "positions": filtered_positions,
-        "velocities": filtered_velocities,
-        "covariances": filtered_covariances,
-        "frame": "UVW",
-        "ephemeris_start": ephemeris_start,
-        "ephemeris_stop": ephemeris_stop,
-        "generated_at": generated_at,
-        "satellite_name": satellite_name,
-        "filename": filename,
-    }
-
-
-def insert_interpolated_splines(interpolated_splines, cursor, connection):
-    """
-    Insert interpolated splines into the database.
-    """
-
-
-def get_starlink_ephemeris_data(cursor, connection):
-    """
-    Fetch and process ephemeris data for Starlink satellites.
-
-    This function:
-    1. Downloads and processes each ephemeris data file
-    2. Parses ephemeris data and inserts it into the database
-
-    The function processes files in the following format:
-    - Each ephemeris file contains satellite position, velocity, and covariance data
-    - Files are named in format:
-        MEME_57851_STARLINK-30405_1490204_Operational_1432778700_UNCLASSIFIED
-
-    Args:
-        connection: Database connection for transaction management
-
-    Raises:
-        requests.HTTPError: If login to Space-Track fails
-        Exception: For other errors during processing
-
-    Note:
-        - Uses ON CONFLICT DO NOTHING for database inserts to handle duplicates
-    """
-    try:
-        with requests.Session() as session:
-            # Retrieve list of ephemeris files
-            ephemeris_response = session.get(
-                "https://api.starlink.com/public-files/ephemerides/MANIFEST.txt",
-                timeout=60,
-            )
-            ephemeris_response.raise_for_status()
-
-            # Parse the manifest to get file names
-            manifest_content = ephemeris_response.text
-            ephemeris_files = [
-                line.strip() for line in manifest_content.splitlines() if line.strip()
-            ]
-
-            files_processed = 0
-            total_data_points = 0
-
-            # Get today's date in UTC
-            today = datetime.now(timezone.utc).date()
-
-            logging.info(f"Today's date: {today}")
-            logging.info(f"Found {len(ephemeris_files)} files in manifest")
-
-            cursor = connection.cursor()
-
-            # Process each file individually
-            for file_name in ephemeris_files:
-                try:
-                    # Check if file has already been processed
-                    cursor.execute(
-                        """
-                        SELECT COUNT(*) FROM interpolable_ephemeris
-                        WHERE file_reference = %s
-                        """,
-                        (file_name,),
-                    )
-                    if cursor.fetchone()[0] > 0:
-                        logging.info(f"Skipping already processed file: {file_name}")
-                        continue
-
-                    # Download the file
-                    file_url = (
-                        f"https://api.starlink.com/public-files/ephemerides/{file_name}"
-                    )
-                    response = session.get(file_url, timeout=120)
-                    response.raise_for_status()
-
-                    logging.info(f"Processing file: {file_name}")
-
-                    try:
-                        file_content = response.content.decode("utf-8")
-                        parsed_data = parse_ephemeris_file(file_content, file_name)
-                        ephemeris, ephemeris_id = insert_ephemeris_data(
-                            parsed_data, cursor, connection
-                        )
-                        # calculate splines and store in database
-                        # ephemeris_data = parse_ephemeris_file(ephemeris_file)
-                        sigma_points_dict = generate_and_propagate_sigma_points(
-                            ephemeris
-                        )
-                        interpolated_splines = interpolate_sigma_pointsKI(  # noqa: F841
-                            sigma_points_dict
-                        )
-
-                        insert_interpolated_splines(
-                            ephemeris_id,
-                            ephemeris.generated_at,
-                            ephemeris.data_source,
-                            ephemeris.frame,
-                            ephemeris.ephemeris_start,
-                            ephemeris.ephemeris_stop,
-                            interpolated_splines,
-                            cursor,
-                            connection,
-                        )
-                        num_points = len(parsed_data["timestamps"]) if ephemeris else 0
-
-                    except Exception as e:
-                        logging.error(
-                            "Error parsing/inserting ephemeris data for file %s: %s",
-                            file_name,
-                            e,
-                        )
-                        try:
-                            connection.rollback()
-                        except Exception as rollback_err:
-                            logging.warning(
-                                "Rollback failed after insert error for %s: %s",
-                                file_name,
-                                rollback_err,
-                            )
-                        cursor = connection.cursor()
-                        continue
-
-                except Exception as e:
-                    logging.error(f"Error processing file {file_name}: {e}")
-                    try:
-                        connection.rollback()
-                    except Exception as rollback_err:
-                        logging.warning(
-                            "Rollback failed after processing error for %s: %s",
-                            file_name,
-                            rollback_err,
-                        )
-                    cursor = connection.cursor()
-                    continue
-
-                total_data_points += num_points
-                files_processed += 1
-
-            logging.info(f"Total data points: {total_data_points}")
-            logging.info(f"Files processed: {files_processed}")
-    except Exception as err:
-        logging.error(f"Error getting ephemeris data for Starlink satellites: {err}")
-        raise
