+++ tests/integration/test_fov_tasks.py
@@ -12,8 +12,10 @@ import numpy as np
 from tests.factories.satellite_factory import SatelliteFactory
 from tests.factories.tle_factory import TLEFactory
 
+from api.adapters.repositories.ephemeris_repository import SqlAlchemyEphemerisRepository
 from api.adapters.repositories.tle_repository import SqlAlchemyTLERepository
 from api.services.tasks.fov_tasks import get_fov_task_status, refine_with_ephemeris_task
+from api.utils.output_utils import format_date
 
 
 def test_get_fov_task_status_pending(app, mocker):
@@ -40,7 +42,14 @@ def test_fov_endpoint_async_integration(client, session, services_available):
     )
     tle = TLEFactory(
         satellite=satellite,
+        tle_line1=(
+            "1 25544U 98067A   24275.50000000  .00012769  00000+0  22936-3 0  9997"
+        ),
+        tle_line2=(
+            "2 25544  51.6416 290.4299 0005730  30.7454 132.9751 15.50238117414255"
+        ),
         epoch=datetime(2024, 10, 1, 0, 0, 0, tzinfo=timezone.utc),
+        data_source="spacetrack",
     )
 
     tle_repo = SqlAlchemyTLERepository(session)
@@ -360,7 +369,7 @@ def test_refine_with_ephemeris_task(mocker):
     fake_session.close.assert_called_once()
 
 
-def test_refine_with_ephemeris_task_keeps_tle_when_krogh_positions_out_of_fov(mocker):
+def test_refine_with_ephemeris_task_position_outside_fov(mocker):
     aggregate_result = (
         [
             {"norad_id": 12345, "orbital_data_source": "tle", "ra": 100.0, "dec": 50.0},
@@ -432,3 +441,68 @@ def test_refine_with_ephemeris_task_keeps_tle_when_krogh_positions_out_of_fov(mo
     assert points == 1
     assert metrics["points_in_fov"] == 1
     fake_session.close.assert_called_once()
+
+
+def test_fov_ephemeris_refinement_end_to_end_real_values(
+    client, session, services_available, starlink_ephemeris
+):
+    """End-to-end check: same scenario with TLE-only vs ephemeris refinement."""
+    satellite = SatelliteFactory(
+        sat_name="STARLINK-31570",
+        sat_number=59324,
+        launch_date=datetime(2020, 1, 1, tzinfo=timezone.utc),
+        decay_date=None,
+    )
+    tle = TLEFactory(
+        satellite=satellite,
+        tle_line1=(
+            "1 59324C 24057E   25283.08937500 -.00019866  00000+0 -71932-3 0  2835"
+        ),
+        tle_line2=(
+            "2 59324  43.0038 317.9750 0001389 281.8907  26.1850 15.27608592    19"
+        ),
+        epoch=datetime(2025, 10, 10, 2, 8, 42, 123456, tzinfo=timezone.utc),
+        data_source="spacetrack",
+    )
+    SqlAlchemyTLERepository(session).add(tle)
+
+    ephemeris_repo = SqlAlchemyEphemerisRepository(session)
+    ephemeris = starlink_ephemeris
+    ephemeris_repo.add(ephemeris)
+    session.commit()
+
+    common_query = (
+        "/fov/satellite-passes/?latitude=0&longitude=0&elevation=0"
+        "&mid_obs_time_jd=2460959.317847&duration=30&ra=312.6525015&dec=-0.9114941"
+        "&fov_radius=180&group_by=satellite&async=false"
+    )
+    tle_only_response = client.get(f"{common_query}&tle_only=true")
+    refined_response = client.get(f"{common_query}&tle_only=false")
+
+    assert tle_only_response.status_code == 200
+    assert refined_response.status_code == 200
+
+    tle_satellites = tle_only_response.json["data"]["satellites"]
+    refined_satellites = refined_response.json["data"]["satellites"]
+
+    tle_matches = [s for s in tle_satellites.values() if s["norad_id"] == 59324]
+    refined_matches = [s for s in refined_satellites.values() if s["norad_id"] == 59324]
+
+    assert tle_matches, "Expected NORAD 59324 in TLE-only response."
+    assert refined_matches, "Expected NORAD 59324 in refined response."
+
+    tle_starlink = tle_matches[0]
+    refined_starlink = refined_matches[0]
+    tle_positions = tle_starlink["positions"]
+    refined_positions = refined_starlink["positions"]
+
+    assert refined_positions[0]["ra"] == 290.30728478
+    assert refined_positions[0]["dec"] == -18.57246244
+    assert tle_positions[0]["ra"] == 290.27460934
+    assert tle_positions[0]["dec"] == -18.5859209
+    assert all(p["orbital_data_source"] == "tle" for p in tle_positions)
+    assert all(p["orbital_data_source"] == "ephemeris" for p in refined_positions)
+    assert all(
+        p["orbital_data_epoch"] == format_date(ephemeris.generated_at)
+        for p in refined_positions
+    )
